feat: refactor ProjectTable into modular components
- Extract TableHeader component for title and header row display
- Create TableTimeAxis component for time axis visualization
- Move ProjectRow component to separate file for better organization
- Extract TimelineBar component with activity visualization logic
- Add SummaryStatistics component for project statistics display
- Create tableUtils for shared calculation functions
- Improve code maintainability and reusability
- Maintain all existing functionality and styling
